+++ core/affiliate/__init__.py
@@ -1,0 +1,5 @@
+"""Affiliate and ad-creative orchestration helpers."""
+
+from .local_higgsfield import LocalHiggsfieldAdapter
+
+__all__ = ["LocalHiggsfieldAdapter"]